Seperate action handler from ActionList
All action logic here remains the same
